requirements.txt
fastapi==0.114.*
uvicorn==0.30.*
faiss-cpu==1.12.0        # <-- bump this
sentence-transformers==3.0.*
numpy==1.26.*
pydantic==2.*
neo4j==5.*
python-dotenv==1.0.*
openai==1.108.1